doc: add examples and improve README
This commit adds examples to the jsonconv package and improves the README file.

The examples demonstrate how to use the jsonconv package to serialize and deserialize Go values to and from JSON. The README file has been updated to include the following:

A description of the package and its features.
Instructions on how to install the package.
Examples of how to use the package.
A list of the functions provided by the package.
A description of the license under which the package is distributed.
These changes make the jsonconv package more user-friendly and easier to understand.